+++ tests/inverse-dynamics/centroidal-id.cpp
@@ -117,15 +117,15 @@ BOOST_AUTO_TEST_CASE(CentroidalID_postureTask)
 
   // Easy access
   const RobotModelHandler & model_handler = test.model_handler;
-  const size_t nq = model_handler.getModel().nq;
-  const size_t nv = model_handler.getModel().nv;
+  const int nq = model_handler.getModel().nq;
+  const int nv = model_handler.getModel().nv;
 
   // Target state
   const Eigen::VectorXd q_target =
     model_handler.getReferenceState().head(nq); // CentroidalID set posture target to reference configuration
   CentroidalID::FeetPoseVector feet_pose_vec;
   CentroidalID::FeetVelocityVector feet_vel_vec;
-  for (int foot_nb; foot_nb < test.model_handler.getFeetNb(); foot_nb++)
+  for (size_t foot_nb; foot_nb < test.model_handler.getFeetNb(); foot_nb++)
   {
     feet_pose_vec.push_back(pinocchio::SE3::Identity());
     feet_vel_vec.push_back(pinocchio::Motion::Zero());
@@ -156,8 +156,8 @@ void test_contact(TestCentroidalID & test)
   // Easy access
   const RobotModelHandler & model_handler = test.model_handler;
   const RobotDataHandler & data_handler = test.data_handler;
-  const size_t nq = model_handler.getModel().nq;
-  const size_t nv = model_handler.getModel().nv;
+  const int nq = model_handler.getModel().nq;
+  const int nv = model_handler.getModel().nv;
 
   // No need to set target as CentroidalID sets it by default to reference state
   const Eigen::VectorXd q_target = model_handler.getReferenceState().head(nq);
@@ -170,7 +170,7 @@ void test_contact(TestCentroidalID & test)
     test.step();
 
     // Check that contact velocity is null
-    for (int foot_nb = 0; foot_nb < model_handler.getFeetNb(); foot_nb++)
+    for (size_t foot_nb = 0; foot_nb < model_handler.getFeetNb(); foot_nb++)
     {
       const pinocchio::Motion foot_vel = pinocchio::getFrameVelocity(
         model_handler.getModel(), data_handler.getData(), model_handler.getFootFrameId(foot_nb), pinocchio::WORLD);
@@ -260,7 +260,7 @@ BOOST_AUTO_TEST_CASE(CentroidalID_baseTask)
 
   // Easy access
   const RobotModelHandler & model_handler = test.model_handler;
-  const size_t nq = model_handler.getModel().nq;
+  const int nq = model_handler.getModel().nq;
 
   // CentroidalID sets posture task by default to reference state
   const Eigen::VectorXd q_target = model_handler.getReferenceState().head(nq);
@@ -424,8 +424,8 @@ BOOST_AUTO_TEST_CASE(CentroidalID_allTasks)
 
   // Easy access
   const RobotModelHandler & model_handler = test.model_handler;
-  const size_t nq = model_handler.getModel().nq;
-  const size_t nv = model_handler.getModel().nv;
+  const int nq = model_handler.getModel().nq;
+  const int nv = model_handler.getModel().nv;
 
   // No need to set target as CentroidalID sets it by default to reference state
   const Eigen::VectorXd q_target = model_handler.getReferenceState().head(nq);
